requirements.txt
streamlit>=1.20.0
ultralytics>=8.0.0
pandas
numpy
Pillow
PyYAML
